requirements.txt
click >=7
ocrd >= 2.0.0a1
pyyaml
requests
